Frontend: fix #4
The issue was caused because the routes/components subscribing to
AuthenticationStatus was not unsubscribing when they were being
destroyed. This was causing the callback given to subscribe method being
fired even when it wasn't necessary. In routes that doesn't allow the
user to be logged in, the callback was redirecting to homepage. As the
unsubscribe function wasn't called, after logging in the callback would
get called whenever the AuthenticationStatus store was getting updated.
This caused the weird issue where the webpage would redirect to `/`
after every 5 seconds.